backend-example/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
# Adicione aqui as dependências do seu modelo de ML:
# tensorflow==2.15.0
# torch==2.1.0
# numpy==1.24.3
# scikit-learn==1.3.2
